// src/escrow/enums/notification-type.enum.ts
export enum NotificationType {
  ESCROW_CREATED = "escrow_created",
  FUNDING_REQUIRED = "funding_required",
  ESCROW_FUNDED = "escrow_funded",
  MILESTONE_STARTED = "milestone_started",
  MILESTONE_COMPLETED = "milestone_completed",
  MILESTONE_APPROVED = "milestone_approved",
  MILESTONE_REJECTED = "milestone_rejected",
  FUNDS_RELEASED = "funds_released",
  ESCROW_COMPLETED = "escrow_completed",
  ESCROW_CANCELLED = "escrow_cancelled",
  DISPUTE_RAISED = "dispute_raised",
  DISPUTE_RESOLVED = "dispute_resolved",
  DEADLINE_APPROACHING = "deadline_approaching",
  DEADLINE_MISSED = "deadline_missed",
  CONDITION_MET = "condition_met",
  CONDITION_FAILED = "condition_failed",
}